Final_Project_n2/requirements_all.txt
# database requirements
psycopg2-binary==2.9.10
Faker==19.3.0
pandas==2.2.2

# airflow dependencies
apache-airflow==2.9.1


